重构：对齐 Android API 契约并修复上传 MIME 识别
- Auth/User/Sync 调用层从 Map 请求体收敛为明确 DTO 模型
- messages/count 响应类型对齐为 Long，sync/push 对齐为 SyncPushRequest
- FileRepositoryImpl 按文件后缀识别真实 MIME，避免统一 octet-stream 导致上传被后端拦截
- 补充并更新相关 repository 测试

验证：./gradlew assembleDebug test

+++ app/src/main/java/com/flashnote/java/data/remote/MessageService.java
@@ -36,5 +36,5 @@ public interface MessageService {
     Call<ApiResponse<Void>> clearInbox();
 
     @GET("api/messages/count")
-    Call<ApiResponse<Integer>> countMessages();
+    Call<ApiResponse<Long>> countMessages();
 }
